WIP - getting more tests to run
Warning: to make KAS work, needed to add underscore.js as a dependency in a script tag
in index.html. Need to re-structure project to not use underscore, or cerate an NPM package
for KAS to include in the project so it can be available during unit tests and prod.