Fix bug with display cleanup of dates in auto-saved files
Auto-save docs have dates and times. To keep them unique,
especially for untitled documents, and because each keystroke
causes an auto-save currently milliseconds are included in the
saved names to keep them unique while a new save is being swapped
out for an old one. Previously these were being inconsistently
generated with 1 or 2 digits in the seconds component. The regex
used to strip off the millis searched for the seconds and miliseconds
to remove them sensitive to the number of digits in the seconds component.